Few enhancements for imageviewer
Control panel has load image at the top now and has a close button.
Full screen toggle now updates image based on if app is full screen or windowed.
When area is selected the context menu can be closed via esc.

Image buttons can have images be dynamically changed by just changing the filename and are lazy loaded.

+++ ZFramework/ZFont.cpp
@@ -1198,7 +1198,7 @@ bool ZFont::Init(const ZFontParams& params)
     }
 
 
-    int createFontHeight = mFontHeight;
+    int createFontHeight = (int)mFontHeight;
     if (UseDoubleSizeFont())
         createFontHeight *= 2;
 
@@ -1456,7 +1456,7 @@ bool ZFont::GenerateGlyph(uint8_t c)
         return true;
 
 
-    DWORD overhang = mFontHeight; // indent by a significant amount to allow space for left overhang
+    DWORD overhang = (DWORD)mFontHeight; // indent by a significant amount to allow space for left overhang
     RECT r;
     r.left = overhang;
     r.top = 0;
